Refactor packages to WellNus fully
We still use the seedu.duke package.

However we've already decided to brand our
  project WellNus, so update the package and
  our classes accordingly.
