--- src/main/java/seedu/duke/atomichabit/command/CommandResult.java
@@ -1,25 +1,0 @@
-package seedu.duke.atomichabit.command;
-
-public class CommandResult {
-    private final String feedbackToUser;
-
-    /**
-     * Constructor to be executed when class is initialized.
-     *
-     * @param feedbackToUser feedback to be printed to user
-     */
-    public CommandResult(String feedbackToUser) {
-        this.feedbackToUser = feedbackToUser;
-    }
-
-    /**
-     * Retrieves the private String feedbackToUser.
-     *
-     * @return String to be printed as output to user
-     */
-    public String getCommandResult() {
-        return feedbackToUser;
-    }
-}
-
-
