--- src/main/java/seedu/duke/exception/AtomicHabitException.java
@@ -1,7 +1,0 @@
-package seedu.duke.exception;
-
-public class AtomicHabitException extends WellNusException {
-    public AtomicHabitException(String errorMessage) {
-        super(errorMessage);
-    }
-}
