--- src/main/java/seedu/duke/exception/BadCommandException.java
@@ -1,7 +1,0 @@
-package seedu.duke.exception;
-
-public class BadCommandException extends WellNusException {
-    public BadCommandException(String errorMessage) {
-        super(errorMessage);
-    }
-}
