--- src/main/java/seedu/duke/exception/WellNusException.java
@@ -1,7 +1,0 @@
-package seedu.duke.exception;
-
-public class WellNusException extends Exception {
-    public WellNusException(String errorMessage) {
-        super(errorMessage);
-    }
-}
