+++ src/main/java/wellnus/exception/AtomicHabitException.java
@@ -1,0 +1,7 @@
+package wellnus.exception;
+
+public class AtomicHabitException extends WellNusException {
+    public AtomicHabitException(String errorMessage) {
+        super(errorMessage);
+    }
+}
