+++ src/main/java/wellnus/exception/BadCommandException.java
@@ -1,0 +1,7 @@
+package wellnus.exception;
+
+public class BadCommandException extends WellNusException {
+    public BadCommandException(String errorMessage) {
+        super(errorMessage);
+    }
+}
