+++ src/main/java/wellnus/exception/WellNusException.java
@@ -1,0 +1,7 @@
+package wellnus.exception;
+
+public class WellNusException extends Exception {
+    public WellNusException(String errorMessage) {
+        super(errorMessage);
+    }
+}
